[Fixes 162] improve rollback procedure and fix bug for cloning

+++ importer/handlers/common/vector.py
@@ -1,5 +1,5 @@
 from importer.publisher import DataPublisher
-from importer.utils import call_rollback_function
+from importer.utils import call_rollback_function, find_key_recursively
 from itertools import chain
 import json
 import logging
@@ -635,41 +635,52 @@ class BaseVectorFileHandler(BaseHandler):
 
         for step in reversed_steps:
             normalized_step_name = step.split(".")[-1]
+            istance_name = find_key_recursively(kwargs, "new_dataset_alternate") or args[3]
             if getattr(self, f"_{normalized_step_name}_rollback", None):
                 function = getattr(self, f"_{normalized_step_name}_rollback")
-                function(exec_id, *args, **kwargs)
+                function(exec_id, istance_name, *args, **kwargs)
 
         logger.warning(f"Rollback for execid: {exec_id} resource published was: {args[3]} completed")
 
-    def _import_resource_rollback(self, exec_id, *args, **kwargs):
+    def _import_resource_rollback(self, exec_id, istance_name=None, *args, **kwargs):
         '''
         We use the schema editor directly, because the model itself is not managed
         on creation, but for the delete since we are going to handle, we can use it
         '''
-        logger.info(f"Rollback dynamic model step in progress for execid: {exec_id} resource published was: {args[3]}")
-        name = args[3]
+        name = istance_name or args[3]
+        logger.info(f"Rollback dynamic model step in progress for execid: {exec_id} resource published was: {name}")
         schema = ModelSchema.objects.filter(name=name).first()
-        _model_editor = ModelSchemaEditor(initial_model=name, db_name=schema.db_name)
-        _model_editor.drop_table(schema.as_model())
-        schema.delete()
-    
-    def _publish_resource_rollback(self, exec_id, *args, **kwargs):
+        if schema is not None:
+            _model_editor = ModelSchemaEditor(initial_model=name, db_name=schema.db_name)
+            _model_editor.drop_table(schema.as_model())
+            schema.delete()
+
+    def _publish_resource_rollback(self, exec_id, istance_name=None, *args, **kwargs):
         '''
         We delete the resource from geoserver
         '''
-        logger.info(f"Rollback publishing step in progress for execid: {exec_id} resource published was: {args[3]}")
+        logger.info(f"Rollback publishing step in progress for execid: {exec_id} resource published was: {istance_name}")
         exec_object = orchestrator.get_execution_object(exec_id)
         handler_module_path = exec_object.input_params.get("handler_module_path")
         publisher = DataPublisher(handler_module_path=handler_module_path)
-        publisher.delete_resource(args[3])
+        publisher.delete_resource(istance_name)
     
-    def _create_geonode_resource_rollback(self, exec_id, *args, **kwargs):
+    def _create_geonode_resource_rollback(self, exec_id, istance_name=None, *args, **kwargs):
         '''
         The handler will remove the resource from geonode
         '''
-        logger.info(f"Rollback geonode step in progress for execid: {exec_id} resource created was: {args[3]}")
-        resource = ResourceBase.objects.filter(alternate__icontains=args[3])
-        resource.delete()
+        name = istance_name or args[3]        
+        logger.info(f"Rollback geonode step in progress for execid: {exec_id} resource created was: {name}")
+        resource = ResourceBase.objects.filter(alternate__icontains=name)
+        if resource.exists():
+            resource.delete()
+    
+    def _copy_dynamic_model_rollback(self, exec_id, istance_name=None, *args, **kwargs):
+        self._import_resource_rollback(exec_id, istance_name=istance_name)
+    
+    def _copy_geonode_resource_rollback(self, exec_id, istance_name=None, *args, **kwargs):
+        self._create_geonode_resource_rollback(exec_id, istance_name=istance_name)
+
 
 
 @importer_app.task(
